# swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters."
  version: "0.0.1"
  title: "Swagger Blog"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost"
basePath: "/v1"
schemes:
  - "https"
  - "http"
paths:
...
  /posts/{postId}/comments/{commentId}:
    put:
      tags:
        - "comment"
      summary: "Updates a comment in the store with  data"
      description: ""
      operationId: "updateComment"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "postId"
          in: "path"
          description: "ID of post that needs to be updated"
          required: true
          type: "integer"
          format: "int64"
        - name: "commentId"
          in: "path"
          description: "Updated commentID of the post"
          required: true
          type: "integer"
          format: "int64"
        - name: "text"
          in: "application/json"
          description: "Updated text of the comment"
          required: false
          type: "string"
      responses:
        405:
          description: "Invalid input"
      security:
        - post_auth:
            - "write:comments"
            - "read:comments"
    delete:
      tags:
        - "comment"
      summary: "Deletes a comment"
      description: ""
      operationId: "deleteComment"
      produces:
        - "application/json"
      parameters:
        - name: "commentId"
          in: "path"
          description: "comment id to delete"
          required: true
          type: "integer"
          format: "int64"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Post not found"
...